restructure src, utils
