+++ streamlit_app.py
@@ -74,8 +74,8 @@ from langchain_community.embeddings import OllamaEmbeddings
 
 
 
-from src.rag.chain import rag_chain_constructor, construct_time_filter
-from src.utils import set_emb_llm
+from utils.rag.chain import rag_chain_constructor, construct_time_filter
+from utils import set_emb_llm
 
 
 from collections import defaultdict
